+++ src/pages/_app.tsx
@@ -4,10 +4,10 @@ import { AppProps } from 'next/app';
 import Head from 'next/head';
 import * as React from 'react';
 import SEO from '../../next-seo.config';
-import { Footer } from '../components/footer/Footer';
 import { mdxComponents } from '../components/mdx/MdxComponents';
-import { NavAdjustedContent } from '../components/navbar/nav.styles';
-import { Navbar } from '../components/navbar/Navbar';
+import { Footer } from '../components/site/footer/Footer';
+import { NavAdjustedContent } from '../components/site/navbar/nav.styles';
+import { Navbar } from '../components/site/navbar/Navbar';
 import { MarginLarge } from '../components/ui-system/margins/MarginLarge';
 import { ContentProvider } from '../context/ContentContext';
 import { ThemeProvider } from '../context/ThemeContext';
